refactor(core): 优化代码逻辑和性能
- 使用新的链式条件语法简化多个嵌套的 if let 表达式
- 替换过时的 map_or 方法为更现代的 is_none_or 和 is_some_and 方法
- 将除法运算替换为更精确的 div_ceil 和 clamp 方法
- 简化量化计算中的向上取整逻辑
- 优化 MIDI 音轨通道值的范围限制逻辑
- 重构代码结构以提高可读性并减少不必要的大括号块
- 更新主题配置以适配 macOS 平台特定的样式设置
- 移除不再使用的播放起始刻度变量和相关方法定义

+++ crates/yinhe-egui/src/time_format.rs
@@ -20,7 +20,7 @@ pub fn format_time_sig(numerator: u8, denominator_power: u8) -> String {
 
 /// Convert tick to `bar.beat.tick_in_beat` format (all 1-indexed).
 pub fn format_tick_bar_beat(tick: f64, ppq: u32, numerator: u8) -> String {
-    let ticks_per_bar = ppq as u32 * numerator as u32;
+    let ticks_per_bar = ppq * numerator as u32;
     let bar = (tick / ticks_per_bar as f64).floor() as u32 + 1;
     let beat = ((tick % ticks_per_bar as f64) / ppq as f64).floor() as u32 + 1;
     let tick_in_beat = (tick % ppq as f64) as u32;
